Extract code in utils
